+++ examples/keycode.rs
@@ -1,6 +1,5 @@
 use bevy::prelude::*;
-use bevy_input_sequence::*;
-use std::time::Duration;
+use bevy_input_sequence::prelude::*;
 
 #[derive(Clone, Debug)]
 enum Direction {
@@ -14,7 +13,8 @@ struct MyEvent(Direction);
 fn main() {
     App::new()
         .add_plugins(DefaultPlugins)
-        .add_key_sequence_event::<MyEvent>()
+        .add_plugins(InputSequencePlugin::default())
+        .add_event::<MyEvent>()
         .add_systems(Startup, setup)
         .add_systems(Update, event_listener)
         .run();
@@ -23,9 +23,9 @@ fn main() {
 #[rustfmt::skip]
 fn setup(mut commands: Commands) {
     // Specify key codes directly.
-    commands.spawn(
+    commands.add(
         KeySequence::new(
-            MyEvent(Direction::Clockwise),
+            action::send_event(MyEvent(Direction::Clockwise)),
             [KeyCode::KeyW,
              KeyCode::KeyD,
              KeyCode::KeyS,
@@ -35,9 +35,9 @@ fn setup(mut commands: Commands) {
     );
 
     // Use keyseq! macro.
-    commands.spawn(
+    commands.add(
         KeySequence::new(
-            MyEvent(Direction::CounterClockwise),
+            action::send_event(MyEvent(Direction::CounterClockwise)),
             keyseq!(W A S D),
         )
         .time_limit(Duration::from_secs(1)),