+++ examples/multiple_input.rs
@@ -1,43 +1,28 @@
-use std::time::Duration;
-
-use bevy::app::{App, Startup, Update};
-use bevy::prelude::{Commands, Event, EventReader, Gamepad, GamepadButtonType};
-use bevy::DefaultPlugins;
-
-use bevy_input_sequence::AddInputSequenceEvent;
-use bevy_input_sequence::{keyseq, ButtonSequence, GamepadEvent, KeySequence};
+use bevy::prelude::*;
+use bevy_input_sequence::prelude::*;
 
 #[derive(Event, Clone, Debug)]
 struct MyEvent(u8, Option<Gamepad>);
 
-impl GamepadEvent for MyEvent {
-    fn gamepad(&self) -> Option<Gamepad> {
-        self.1
-    }
-
-    fn set_gamepad(&mut self, gamepad: Gamepad) {
-        self.1 = Some(gamepad);
-    }
-}
-
 fn main() {
     App::new()
         .add_plugins(DefaultPlugins)
-        .add_key_sequence_event::<MyEvent>()
-        .add_button_sequence_event::<MyEvent>()
+        .add_plugins(InputSequencePlugin::default())
+        .add_event::<MyEvent>()
         .add_systems(Startup, setup)
         .add_systems(Update, input_sequence_event_system)
         .run();
 }
 
 fn setup(mut commands: Commands) {
-    commands.spawn(
-        KeySequence::new(MyEvent(1, None), keyseq!(W D S A)).time_limit(Duration::from_secs(5)),
+    commands.add(
+        KeySequence::new(action::send_event(MyEvent(1, None)), keyseq!(W D S A))
+            .time_limit(Duration::from_secs(5)),
     );
 
-    commands.spawn(
+    commands.add(
         ButtonSequence::new(
-            MyEvent(2, None),
+            action::send_event_with_input(|gamepad| MyEvent(2, Some(gamepad))),
             [
                 GamepadButtonType::North,
                 GamepadButtonType::East,
@@ -48,13 +33,14 @@ fn setup(mut commands: Commands) {
         .time_limit(Duration::from_secs(5)),
     );
 
-    commands.spawn(
-        KeySequence::new(MyEvent(3, None), keyseq!(W A S D)).time_limit(Duration::from_secs(5)),
+    commands.add(
+        KeySequence::new(action::send_event(MyEvent(3, None)), keyseq!(W A S D))
+            .time_limit(Duration::from_secs(5)),
     );
 
-    commands.spawn(
+    commands.add(
         ButtonSequence::new(
-            MyEvent(4, None),
+            action::send_event_with_input(|gamepad| MyEvent(4, Some(gamepad))),
             [
                 GamepadButtonType::North,
                 GamepadButtonType::West,
@@ -71,7 +57,10 @@ fn setup(mut commands: Commands) {
 
 fn input_sequence_event_system(mut er: EventReader<MyEvent>) {
     for e in er.read() {
-        println!("{e:?} emitted {}", e.gamepad()
-                 .map(|x| format!("from gamepad id {}", x.id)).unwrap_or("not from gamepad".into()));
+        println!(
+            "{e:?} emitted {}",
+            e.1.map(|x| format!("from gamepad id {}", x.id))
+                .unwrap_or("not from gamepad".into())
+        );
     }
 }