+++ src/input_sequence.rs
@@ -1,70 +1,125 @@
-use bevy::prelude::{Component, Event, GamepadButtonType};
+//! Input sequences for keys and gamepad buttons
+use crate::{cond_system::IntoCondSystem, time_limit::TimeLimit, KeyChord};
 
-use crate::time_limit::TimeLimit;
-use crate::{GamepadEvent, KeyChord};
+use bevy::{
+    ecs::{
+        component::Component,
+        entity::Entity,
+        system::{IntoSystem, System, SystemId},
+        world::World,
+    },
+    input::gamepad::{Gamepad, GamepadButtonType},
+    reflect::Reflect,
+};
 
-/// An input sequence is a series of acts [A] that fires an event when matched
-/// with inputs within the given time limit.
-#[derive(Component, Debug, Clone)]
-pub struct InputSequence<E, A> {
+/// An input sequence is a series of acts that fires an event when matched with
+/// inputs within the given time limit.
+#[derive(Component, Reflect, Clone)]
+#[reflect(from_reflect = false)]
+pub struct InputSequence<Act, In> {
     /// Event emitted
-    pub event: E,
+    #[reflect(ignore)]
+    pub system_id: SystemId<In>,
     /// Sequence of acts that trigger input sequence
-    pub acts: Vec<A>,
+    pub acts: Vec<Act>,
     /// Optional time limit after first match
     pub time_limit: Option<TimeLimit>,
 }
 
-impl<E, A> InputSequence<E, A>
-where
-    E: Event + Clone,
-    A: Clone,
-{
+/// An input sequence builder.
+pub struct InputSequenceBuilder<Act, S> {
+    /// The action when to run when sequence matches
+    pub system: S,
+    /// Sequence of acts that trigger input sequence
+    pub acts: Vec<Act>,
+    /// Optional time limit after first match
+    pub time_limit: Option<TimeLimit>,
+}
+
+impl<Act> InputSequenceBuilder<Act, ()> {
     /// Create new input sequence. Not operant until added to an entity.
-    #[inline(always)]
-    pub fn new<T>(event: E, acts: impl IntoIterator<Item = T>)
-                  -> InputSequence<E, A>
+    pub fn new<C, I, M>(system: C) -> InputSequenceBuilder<Act, C::System>
     where
-        A: From<T>,
+        C: IntoCondSystem<I, (), M> + 'static,
+        I: Send + Sync + 'static,
     {
-        Self {
-            event,
+        InputSequenceBuilder {
+            acts: Vec::new(),
+            system: IntoSystem::into_system(system),
             time_limit: None,
-            acts: Vec::from_iter(acts.into_iter().map(A::from)),
         }
     }
+}
 
+impl<Act, S> InputSequenceBuilder<Act, S>
+where
+    S: System<Out = ()>,
+{
     /// Specify a time limit from the start of the first matching input.
     pub fn time_limit(mut self, time_limit: impl Into<TimeLimit>) -> Self {
         self.time_limit = Some(time_limit.into());
         self
     }
+
+    /// Build the InputSequence. Requires world to register the system.
+    pub fn build(self, world: &mut World) -> InputSequence<Act, S::In> {
+        InputSequence {
+            system_id: world.register_system(self.system),
+            acts: self.acts,
+            time_limit: self.time_limit,
+        }
+    }
+}
+
+impl<Act, S> bevy::ecs::system::Command for InputSequenceBuilder<Act, S>
+where
+    Act: Send + Sync + 'static,
+    S: System<Out = ()> + Send + Sync + 'static,
+    S::In: Send + Sync + 'static,
+{
+    fn apply(self, world: &mut World) {
+        let act = self.build(world);
+        world.spawn(act);
+    }
+}
+
+impl<Act, S> bevy::ecs::system::EntityCommand for InputSequenceBuilder<Act, S>
+where
+    Act: Send + Sync + 'static,
+    S: System<Out = ()> + Send + Sync + 'static,
+    S::In: Send + Sync + 'static,
+{
+    fn apply(self, id: Entity, world: &mut World) {
+        let act = self.build(world);
+        let mut entity = world.get_entity_mut(id).unwrap();
+        entity.insert(act);
+    }
+}
+
+impl<Act, In> InputSequence<Act, In>
+where
+    In: 'static,
+{
+    /// Create new input sequence. Not operant until added to an entity.
+    #[inline(always)]
+    #[allow(clippy::new_ret_no_self)]
+    pub fn new<T, C, I, M>(
+        system: C,
+        acts: impl IntoIterator<Item = T>,
+    ) -> InputSequenceBuilder<Act, C::System>
+    where
+        C: IntoCondSystem<I, (), M> + 'static,
+        Act: From<T>,
+        I: Send + Sync + 'static,
+    {
+        let mut builder = InputSequenceBuilder::new(system);
+        builder.acts = Vec::from_iter(acts.into_iter().map(Act::from));
+        builder
+    }
 }
 
 /// Represents a key sequence.
-pub type KeySequence<E> = InputSequence<E, KeyChord>;
+pub type KeySequence = InputSequence<KeyChord, ()>;
 
-// pub type ButtonSequence<E> = InputSequence<E, GamepadButtonType>;
 /// Represents a gamepad button sequence.
-#[derive(Component, Debug, Clone)]
-pub struct ButtonSequence<E>(pub(crate) InputSequence<E, GamepadButtonType>);
-
-impl<E> ButtonSequence<E>
-where
-    E: GamepadEvent + Clone,
-{
-    /// Create new button sequence. Not operant until added to an entity.
-    #[inline(always)]
-    pub fn new(event: E, acts: impl IntoIterator<Item = GamepadButtonType>) -> ButtonSequence<E> {
-        ButtonSequence(InputSequence {
-            event,
-            time_limit: None,
-            acts: Vec::from_iter(acts),
-        })
-    }
-    /// Specify a time limit from the start of the first matching input.
-    pub fn time_limit(mut self, time_limit: impl Into<TimeLimit>) -> Self {
-        self.0.time_limit = Some(time_limit.into());
-        self
-    }
-}
+pub type ButtonSequence = InputSequence<GamepadButtonType, Gamepad>;