+++ CHANGELOG.md
@@ -2,13 +2,28 @@
 
 All notable changes to this project will be documented in this file.
 
-## [released]
+## [unreleased]
+
+### Features
+
+- Generalize to running one-shot systems instead of firing events.
+- Use a plugin, can configure schedule and system set.
+- Remove `GamepadEvent`; use system with input `In<Gamepad>`.
+- Add IntoCondSystem that adds `only_if()` conditions to `IntoSystems`.
+- Add `only_if` example.
+- Add prelude module for glob imports.
+
+### Refactor
+
+- Hollow out lib so it's just `mod` and `pub use` statements.
+- Extract sets of functionality from `lib.rs` into `chord.rs`, `cache.rs` and `plugin.rs`.
+- Extract simulated tests into `tests/` directory.
 
 ## [0.3.0] - 2024-03-08
 
 ### Features
 
-- Removed `Act::Any`, so We can no longer define an act composed of multiple buttons(or keycodes).
+- Removed `Act::Any`, so we can no longer define an act composed of multiple buttons(or keycodes).
 - The keyboard and gamepad now use different sequences.